+++ eslint.config.js
@@ -14,7 +14,9 @@ export default [
 			'**/test-results',
 			'**/build',
 			'**/.custom-out-dir',
-			'packages/adapter-*/files'
+			'packages/adapter-*/files',
+			'packages/kit/src/core/config/fixtures/multiple', // dir contains svelte config with multiple extensions tripping eslint
+			'packages/package/test/fixtures/typescript-svelte-config/expected'
 		]
 	},
 	{
@@ -41,7 +43,8 @@ export default [
 			'packages/kit/test/build-errors/**/*',
 			'packages/kit/test/prerendering/**/*',
 			'packages/package/test/errors/**/*',
-			'packages/package/test/fixtures/**/*'
+			'packages/package/test/fixtures/**/*',
+			'packages/test-redirect-importer/index.js'
 		]
 	}
 ];