+++ packages/adapter-cloudflare/ambient.d.ts
@@ -8,6 +8,8 @@ declare global {
 	namespace App {
 		export interface Platform {
 			env: unknown;
+			ctx: ExecutionContext;
+			/** @deprecated Use `ctx` instead */
 			context: ExecutionContext;
 			caches: CacheStorage;
 			cf?: IncomingRequestCfProperties;